feat: multitenancy (#677)
* init web-js recipes, supertokens from auth-react

* move recipe output type to types.ts

* make combined recipes use web-js version

* get rid of recipeImplementation for single recipes

* add functionOverrides to thirdpartypasswordless recipe

* add functionOverrides to thirdpartyemailpassword recipe

* get rid of one the duplicate setLoginAttemptInfo calls

* fix tests

* update changelog

* fixes after review

* move initOutput type to root types

* fix tests

* review fixes

* review fixes

* add build files

* use normalisedConfig for webjs init in combined recipes

* remove commented code

* disable isIE check in order to avoid error

* test fixes caused by utils

* fix typo

* rename recipeImpl -> webJSRecipe for all recipes

* add comments

* utils-> normalization_function  fix

* type changes for recipe member

* type change, renamein

* move config normalization to init for session

* Update thirdpartypasswordless comments


* Update thirdpartyemailpassword/recipe comments


* review fixes

* reuse action type for RecipeInitResult

* move onHandelEvent call to combined recipe overrides

* fixes-session, overrides

* add callbacks to isVerified claim

* fix exampleTestHelper

* fix build folder

* add auth-react boolean claim

* extend web-js claims

* fix tests, rename callbacks

* review fixes

* fix tests

* use web-js import instead of website

* fix type GetRedirectionURLContext, add comment, pass context to getGlobalClaim function calls

* fix duplication

* move helper function to utils

* fix emailverification wrapper type

* self review fixes

* second round review fixed

* add access-denied page

* add ability to override and change style of access-denied screen

* add denial info to access denied page

* update example apps

* fix rollup watch mac issue

* review fixes

* fix redirection priorities

* move cliam functions to session utils

* fix util function

* add unit tests for new utils function

* initial splitting

* create class for each recipe to handle routes

* review fixes for /lib

* update examples

* fix partially sessionAuth tests

* fix: typo

* fix: sessionAuth test

* update changelog

* fix lintiing issues

* fix tests

* Merge branch '0.31' into 'feat/splitting-prebuild-ui-components'

* fix: partially unit test

* fix: unit test case

* fix: broken components overriding after merge

* fix: move themeBase & fix example app

* fix: type fix

* fix: access denial info

* fix: exporting claims from auth-react

* fix: review fixes

* fix: by review comments

* fix: adjust access-denied page

* more review fixes

* fix: review fixes

* change implementataion & access denied page design

* fix: default page & review changes

* fix: default failed claim when none has callback

* css: override container width

* fix: review comments

* fix: add comment

* fix: respect rid when rendering UIs

* fix tests

* fix: add underscore prefix

* test: add test for back button

* fix: emailverification prebuilt ui

* fix: review changes

* fix no router case

* fix: prebuiltui components types exoprt

* fix: emailverification exports

* fix: test apps

* fix: adjust routing component to handle duplicate renders

* feat: clean up onSuccessRedirection and add showAccessDeniedOnFailure

* fix: e2e tests

* update changelog

* fix: example apps

* fix examples & changelog

* fix: preBuiltUI -> prebuiltui

* fix filename casing

* fix: preBuiltUI -> prebuiltui

* fix filename casing

* expose signle routing functons

* rename: prebuiltui->ui

* update test apps

* fix: unit tests

* preBuiltUI -> prebuiltui

* update changelog

* update example apps

* getSuperTokensReactRouterDomRoutes casing fix

* getSuperTokensReactRouterDomRoutes casing fix

* review fixes

* review fixes

* review fixes

* wip

* fix: review comments(preBuiltUI interface consistency)

* parent c3935c9925aed8a9588c2188446a6a938b7de340
author Alisher <alisher@supertokens.com> 1679996212 +0400
committer Alisher <alisher@supertokens.com> 1682072680 +0400

feat: multitenancy

* feat: thirdparty rework (#676)

* add missing providers & add prop to buttonComponent

* fix: Thirdparty login popup (#675)

* fix: thirdparty popup example

* chore: changelog

* fix: adding backend

* fix: pr comment

* fix: pr comment

* fix: pr comment

* fix: pr comment

* feat: add new providers & add prop to buttonComponent

* fix: activeDirectory width

* update changelog

* Update examples/with-thirdparty-popup/README.md


* fix: providers

* fix: move getButton to base class

* fix: review comments

* Update examples/with-thirdparty-popup/frontend/src/App.tsx


* fixes

---------


* Update lib/ts/recipe/multitenancy/index.ts


* fix: review comments

* revert providers form

* fix: review 2nd round

* fixes: getRoutingComponent

* fix: review 2nd round & embedding components fixes

* fix: review

* fix: move provider picking logic into separate wrapper

* fix: remove test data

* fix: picking providers logic & types

* fix: type fixes & move helper to utils

* fix: add getter for dynamicLoginMethods

* fix: provider selection algorithm

* fixes

* fixes

* remove memoization of providders

* fixes

* chore: minor cleanup/test fixes + migration docs

* test: revert merge issue

* fixes

* fixes

* fixes

* fixes

* fixes

* fixes

* fixes reciperouter

* fixes

* fix: update to match ADRs & new BE version

* test: multitenancy (#678)

* wip: tests

* test: additional tests

* fix: picking providers logic & types

* test: adds test case

* fix test & remove redundant test case

* fix: minor cleanup/small fix + test extension

---------


* feat: update to latest interfaces and fix tests

* test: updated test asserts to match new behaviour

* fix: make the interface more consistent with web-js

* feat: improve typing of getClaimValue

* feat: update tests to use core + small fixes

* test: add tests for tenant id exceptions

* fix: fixed handling empty provider arrays + missing provider

* fix: fix tests

* fix: self-review fixes

* fix: self-review fixes & chores

* test: fix react 16 test app

* docs: update examples

---------

+++ test/startTestApp.sh
@@ -11,9 +11,13 @@ export EXIT_PID=$$
 apiPort=$1
 
 function killServers () {
-    echo "Kill servers."
-    lsof -i tcp:8082 | grep -m 1 node | awk '{printf $2}' | cut -c 1- | xargs -I {} kill -9 {} > /dev/null 2>&1
-    lsof -i tcp:3031 | grep -m 1 node | awk '{printf $2}' | cut -c 1- | xargs -I {} kill -9 {} > /dev/null 2>&1
+    if [[ "${SERVER_STARTED}" != "true" ]]; then
+        echo "Kill servers."
+        lsof -i tcp:8082 | grep -m 1 node | awk '{printf $2}' | cut -c 1- | xargs -I {} kill -9 {} > /dev/null 2>&1
+        lsof -i tcp:3031 | grep -m 1 node | awk '{printf $2}' | cut -c 1- | xargs -I {} kill -9 {} > /dev/null 2>&1
+    else
+        echo "Leaving servers running because SERVER_STARTED=true"
+    fi
 }
 
 function startEndToEnd () {
@@ -26,7 +30,10 @@ function startEndToEnd () {
     sleep 2 # Because the server is responding does not mean the app is ready. Let's wait another 5secs to make sure the app is up.
     echo "Start mocha testing"
 
-    if ! [[ -z "${SPEC_FILES}" ]]; then
+    if ! [[ -z "${GREP}" ]]; then
+        echo "$GREP"
+        APP_SERVER=$apiPort TEST_MODE=testing mocha --require @babel/register --require test/test.mocha.env --timeout 40000 --no-config --grep "$GREP"
+    elif ! [[ -z "${SPEC_FILES}" ]]; then
         APP_SERVER=$apiPort TEST_MODE=testing mocha --require @babel/register --require test/test.mocha.env --timeout 40000 --no-config $SPEC_FILES
     elif [[ -z "${MOCHA_FILE}" ]]; then
         APP_SERVER=$apiPort TEST_MODE=testing mocha --require @babel/register --require test/test.mocha.env --timeout 40000 --no-config test/end-to-end/**/*.test.js
@@ -69,9 +76,12 @@ mkdir -p test_report/logs
 
 if [[ "${RUN_REACT_16_TESTS}" == "true" ]]; then
     echo "Running tests with React 16"
-    (cd test/server/ && TEST_MODE=testing INSTALL_PATH=../../../supertokens-root NODE_PORT=8082 node . >> ../../test_report/logs/backend-react16.log 2>&1 &)
-    
-    (cd examples/for-tests-react-16/ && cat | CI=true BROWSER=none PORT=3031 REACT_APP_API_PORT=$apiPort npm run start >> ../../test_report/logs/frontend-react-16.log 2>&1 &)
+
+    if [[ "${SERVER_STARTED}" != "true" ]]; then
+        (cd test/server/ && TEST_MODE=testing INSTALL_PATH=../../../supertokens-root NODE_PORT=8082 node . >> ../../test_report/logs/backend-react16.log 2>&1 &)
+        
+        (cd examples/for-tests-react-16/ && cat | CI=true BROWSER=none PORT=3031 REACT_APP_API_PORT=$apiPort npm run start >> ../../test_report/logs/frontend-react-16.log 2>&1 &)
+    fi
     
     IS_REACT_16=true RUN_RRD5=true startEndToEnd
 
@@ -79,9 +89,11 @@ if [[ "${RUN_REACT_16_TESTS}" == "true" ]]; then
 else
     echo "Running tests with React 18"
     # Run node server in background.
-    (cd test/server/ && TEST_MODE=testing INSTALL_PATH=../../../supertokens-root NODE_PORT=8082 node . >> ../../test_report/logs/backend.log 2>&1 &)
+    if [[ "${SERVER_STARTED}" != "true" ]]; then
+        (cd test/server/ && TEST_MODE=testing INSTALL_PATH=../../../supertokens-root NODE_PORT=8082 node . >> ../../test_report/logs/backend.log 2>&1 &)
 
-    (cd ./examples/for-tests/ && cat | CI=true BROWSER=none PORT=3031 REACT_APP_API_PORT=$apiPort npm run start >> ../../test_report/logs/frontend.log 2>&1 &)
+        (cd ./examples/for-tests/ && cat | CI=true BROWSER=none PORT=3031 REACT_APP_API_PORT=$apiPort npm run start >> ../../test_report/logs/frontend.log 2>&1 &)
+    fi
     # Start front end test app and run tests.
     startEndToEnd
